@@ -1,21 +1,21 @@
 package art
 
-// node constraints
+// node constraints.
 const (
-	node4Min = 2
-	node4Max = 4
+	node4Min = 2 // minimum number of children for node4.
+	node4Max = 4 // maximum number of children for node4.
 
-	node16Min = node4Max + 1
-	node16Max = 16
+	node16Min = node4Max + 1 // minimum number of children for node16.
+	node16Max = 16           // maximum number of children for node16.
 
-	node48Min = node16Max + 1
-	node48Max = 48
+	node48Min = node16Max + 1 // minimum number of children for node48.
+	node48Max = 48            // maximum number of children for node48.
 
-	node256Min = node48Max + 1
-	node256Max = 256
+	node256Min = node48Max + 1 // minimum number of children for node256.
+	node256Max = 256           // maximum number of children for node256.
 )
 
 const (
-	// MaxPrefixLen is maximum prefix length for internal nodes.
-	MaxPrefixLen = 10
+	// maxPrefixLen is maximum prefix length for internal nodes.
+	maxPrefixLen = 10
 )
