@@ -5,39 +5,38 @@
 //
 // Usage
 //
-//  package main
+//	package main
 //
-//  import (
-//     "fmt"
-//     "github.com/plar/go-adaptive-radix-tree"
-//  )
+//	import (
+//	   "fmt"
+//	   "github.com/plar/go-adaptive-radix-tree"
+//	)
 //
-//  func main() {
+//	func main() {
 //
-//     tree := art.New()
+//	   tree := art.New()
 //
-//     tree.Insert(art.Key("Hi, I'm Key"), "Nice to meet you, I'm Value")
-//     value, found := tree.Search(art.Key("Hi, I'm Key"))
-//     if found {
-//         fmt.Printf("Search value=%v\n", value)
-//     }
+//	   tree.Insert(art.Key("Hi, I'm Key"), "Nice to meet you, I'm Value")
+//	   value, found := tree.Search(art.Key("Hi, I'm Key"))
+//	   if found {
+//	       fmt.Printf("Search value=%v\n", value)
+//	   }
 //
-//     tree.ForEach(func(node art.Node) bool {
-//         fmt.Printf("Callback value=%v\n", node.Value())
-//         return true
-//     }
+//	   tree.ForEach(func(node art.Node) bool {
+//	       fmt.Printf("Callback value=%v\n", node.Value())
+//	       return true
+//	   }
 //
-//     for it := tree.Iterator(); it.HasNext(); {
-//         value, _ := it.Next()
-//         fmt.Printf("Iterator value=%v\n", value.Value())
-//     }
-//  }
+//	   for it := tree.Iterator(); it.HasNext(); {
+//	       value, _ := it.Next()
+//	       fmt.Printf("Iterator value=%v\n", value.Value())
+//	   }
+//	}
 //
-//  // Output:
-//  // Search value=Nice to meet you, I'm Value
-//  // Callback value=Nice to meet you, I'm Value
-//  // Iterator value=Nice to meet you, I'm Value
-//
+//	// Output:
+//	// Search value=Nice to meet you, I'm Value
+//	// Callback value=Nice to meet you, I'm Value
+//	// Iterator value=Nice to meet you, I'm Value
 //
 // Also the current implementation was inspired by [2] and [3]
 //
